refactor: SaveAfternoteUseCase 도메인 이전 및 에디터 저장 로직 리팩터링
- `SaveAfternoteUseCase`를 presentation 레이어에서 domain 레이어로 이동하고 `SaveAfternoteCommand`를 사용하는 구조로 개선
- `CreateAfternoteInput`, `SaveAfternoteCommand` 등 도메인 모델을 추가하여 에디터 생성 및 수정 로직 추상화
- `AfternoteEditorViewModel` 내부에 저장 명령 생성 로직(`buildSaveCommand`)을 분리하고 새 도메인 UseCase 반영
- `AfternoteEditorFormMapper`의 반환 타입을 도메인 모델로 변경하고 내부적으로 정의되었던 `CreateInput` 인터페이스 제거
- 미사용 `ReceiverAfternoteTriggerViewModel` 및 관련 이벤트 정의 파일 삭제
- Dagger/Hilt 주입 및 리소스 어노테이션에 `@param:` 접두사를 추가하여 어노테이션 타겟 최적화 (Kotlin 대응)
