+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/author/SaveAfternoteCommand.kt
@@ -1,0 +1,12 @@
+package com.afternote.feature.afternote.domain.model.author
+
+sealed interface SaveAfternoteCommand {
+    data class Create(
+        val input: CreateAfternoteInput,
+    ) : SaveAfternoteCommand
+
+    data class Update(
+        val id: Long,
+        val payload: AfternoteUpdatePayload,
+    ) : SaveAfternoteCommand
+}
