+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/AfternoteEditorViewModel.kt
@@ -8,9 +8,12 @@ import com.afternote.core.model.AlbumCover
 import com.afternote.feature.afternote.domain.error.AfternoteAuthoringValidationException
 import com.afternote.feature.afternote.domain.error.AfternoteAuthoringValidationKind
 import com.afternote.feature.afternote.domain.model.author.AuthorReceiverEntry
+import com.afternote.feature.afternote.domain.model.author.SaveAfternoteCommand
 import com.afternote.feature.afternote.domain.repository.author.AfternoteRepository
 import com.afternote.feature.afternote.domain.repository.author.AuthorReceiverRepository
 import com.afternote.feature.afternote.domain.repository.author.MemorialThumbnailUploadRepository
+import com.afternote.feature.afternote.domain.usecase.editor.ResolveMemorialMediaForSaveUseCase
+import com.afternote.feature.afternote.domain.usecase.editor.SaveAfternoteUseCase
 import com.afternote.feature.afternote.presentation.R
 import com.afternote.feature.afternote.presentation.author.editor.mapper.toAfternoteEditorReceivers
 import com.afternote.feature.afternote.presentation.author.editor.memorial.playlist.Song
@@ -212,6 +215,7 @@ class AfternoteEditorViewModel
         private val afternoteRepository: AfternoteRepository,
         private val memorialThumbnailUploadRepository: MemorialThumbnailUploadRepository,
         private val saveAfternoteUseCase: SaveAfternoteUseCase,
+        private val resolveMemorialMediaForSave: ResolveMemorialMediaForSaveUseCase,
     ) : ViewModel() {
         private val formSnapshotJson =
             Json {
@@ -344,7 +348,7 @@ class AfternoteEditorViewModel
                 internalState.update {
                     it.copy(isSaving = true, error = null, errorRes = null, validationError = null)
                 }
-                saveAfternoteUseCase(
+                buildSaveCommand(
                     editingId = editingId,
                     categoryForApi = categoryForApi,
                     payload = payload,
@@ -352,16 +356,70 @@ class AfternoteEditorViewModel
                     playlistSongs = playlistSongs,
                     memorialMedia = memorialMedia,
                 ).fold(
-                    onSuccess = { id ->
-                        Log.d(TAG, "saveAfternote: SUCCESS, savedId=$id")
-                        internalState.update {
-                            it.copy(isSaving = false, savedId = id)
-                        }
-                        _events.send(AfternoteEditorEvent.SaveSuccess(id))
+                    onSuccess = { command ->
+                        saveAfternoteUseCase(command).fold(
+                            onSuccess = { id ->
+                                Log.d(TAG, "saveAfternote: SUCCESS, savedId=$id")
+                                internalState.update {
+                                    it.copy(isSaving = false, savedId = id)
+                                }
+                                _events.send(AfternoteEditorEvent.SaveSuccess(id))
+                            },
+                            onFailure = { e -> handleSaveFailure(e, categoryForApi) },
+                        )
                     },
                     onFailure = { e -> handleSaveFailure(e, categoryForApi) },
                 )
             }
+        }
+
+        private suspend fun buildSaveCommand(
+            editingId: Long?,
+            categoryForApi: EditorCategory,
+            payload: RegisterAfternotePayload,
+            selectedReceiverIds: List<Long>,
+            playlistSongs: List<Song>,
+            memorialMedia: SaveAfternoteMemorialMedia,
+        ): Result<SaveAfternoteCommand> {
+            val resolved =
+                resolveMemorialMediaForSave(
+                    funeralVideoUrl = memorialMedia.funeralVideoUrl,
+                    memorialPhotoUrl = memorialMedia.memorialPhotoUrl,
+                    pickedMemorialPhotoUri = memorialMedia.pickedMemorialPhotoUri,
+                    funeralThumbnailUrl = memorialMedia.funeralThumbnailUrl,
+                    isUpdate = editingId != null,
+                ).getOrElse { return Result.failure(it) }
+
+            val command =
+                if (editingId != null) {
+                    val updatePayload =
+                        AfternoteEditorFormMapper.buildUpdatePayload(
+                            category = categoryForApi,
+                            payload = payload,
+                            selectedReceiverIds = selectedReceiverIds,
+                            playlistSongs = playlistSongs,
+                            memorialMedia =
+                                MemorialMediaUrls(
+                                    funeralVideoUrl = resolved.videoUrlForUpdate,
+                                    funeralThumbnailUrl = resolved.funeralThumbnailUrlForUpdate,
+                                    memorialPhotoUrl = resolved.resolvedMemorialPhotoUrl,
+                                ),
+                        )
+                    SaveAfternoteCommand.Update(id = editingId, payload = updatePayload)
+                } else {
+                    val createInput =
+                        AfternoteEditorFormMapper.buildCreateInput(
+                            category = categoryForApi,
+                            payload = payload,
+                            selectedReceiverIds = selectedReceiverIds,
+                            playlistSongs = playlistSongs,
+                            funeralVideoUrl = resolved.resolvedVideoUrl,
+                            funeralThumbnailUrl = memorialMedia.funeralThumbnailUrl,
+                            memorialPhotoUrl = resolved.resolvedMemorialPhotoUrl,
+                        )
+                    SaveAfternoteCommand.Create(input = createInput)
+                }
+            return Result.success(command)
         }
 
         private fun loadExistingAfternoteForEdit(afternoteId: Long) {
